backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
python-dotenv